+++ src/renderer/features/integrations/communications-store.ts
@@ -8,6 +8,8 @@
 
 import { create } from 'zustand';
 
+import { loadRules, saveRules } from './notification-rules-storage';
+
 // ── Types ────────────────────────────────────────────────────
 
 type ServiceStatus = 'connected' | 'disconnected' | 'error';
@@ -42,34 +44,6 @@ interface IntegrationsState {
   addNotificationRule: (rule: Omit<NotificationRule, 'id'>) => void;
   removeNotificationRule: (id: string) => void;
   toggleNotificationRule: (id: string) => void;
-}
-
-// ── localStorage helpers ──────────────────────────────────────
-
-const RULES_STORAGE_KEY = 'adc:notification-rules';
-
-function loadRules(): NotificationRule[] {
-  try {
-    const raw = localStorage.getItem(RULES_STORAGE_KEY);
-    if (raw === null) return [];
-    const parsed: unknown = JSON.parse(raw);
-    if (!Array.isArray(parsed)) return [];
-    return parsed.filter(
-      (item): item is NotificationRule =>
-        typeof item === 'object' &&
-        item !== null &&
-        typeof (item as Record<string, unknown>).id === 'string' &&
-        typeof (item as Record<string, unknown>).service === 'string' &&
-        typeof (item as Record<string, unknown>).pattern === 'string' &&
-        typeof (item as Record<string, unknown>).enabled === 'boolean',
-    );
-  } catch {
-    return [];
-  }
-}
-
-function saveRules(rules: NotificationRule[]): void {
-  localStorage.setItem(RULES_STORAGE_KEY, JSON.stringify(rules));
 }
 
 // ── Store ────────────────────────────────────────────────────
